Refacto MVVM : PlayerController + corrections asset playback
  Refactorisation majeure de l'architecture vers MVVM avec séparation
  claire des responsabilités entre UI, ViewModel et Service.

  ## Architecture
  - Introduction de la couche Service avec coordinateurs de playback
  - Création des ViewModels avec JavaFX Properties pour binding UI
  - Séparation des interfaces Playback (AssetPlayback / TimelinePlayback)
  - Suppression de JavaCvPlaybackCoordinator (logique déplacée en Service)

  ## PlayerController (MVVM)
  - Refacto complet : binding UI ↔ PlaybackViewModel
  - Gestion intelligente du drag du slider (flag userIsDragging)
  - Bindings pour activation/désactivation des contrôles
  - Suppression des méthodes publiques (setDuration, enableControls)
  - Thread-safety : Platform.runLater dans les coordinateurs

  ## Corrections asset playback
  - Fix seek : standardisation des unités en millisecondes partout
    (VideoDecoder.seek accepte maintenant long ms au lieu de double seconds)
  - Fix fin de lecture : ajout player.pause() dans callback onEndOfMedia
    pour éviter la lecture en boucle
  - Suppression du double appel onEndOfMedia (videoLoop)
  - Nettoyage des System.out.println de debug

  ## PlaybackViewModel
  - Ajout de propriétés : playbackState, currentAsset, ready
  - Propriété calculée formattedTime (binding automatique)
  - Support pour volume et playbackSpeed (préparation future)

  ## Modèles
  - Conversion de plusieurs classes en records immuables
    (MediaData, MediaInfo, MediaStream, frames)

+++ src/main/java/io/onelioh/babycut/engine/decoder/VideoSeekMarker.java
@@ -3,11 +3,11 @@ package io.onelioh.babycut.engine.decoder;
 import org.bytedeco.javacv.Frame;
 
 public class VideoSeekMarker extends VideoFrame {
-    public VideoSeekMarker(double timestampSeconds) {
+    public VideoSeekMarker(long timestampSeconds) {
         super(timestampSeconds);
     }
 
-    public VideoSeekMarker(Frame frame, double timestampSeconds) {
+    public VideoSeekMarker(Frame frame, long timestampSeconds) {
         super(frame, timestampSeconds);
     }
 }
